pyproject.toml
[project]
name = "heurist-agent-framework"
version = "0.1.0"
description = "Heurist Agent Framework"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "aiohttp==3.11.15",
    "apify-client==1.9.2", # truth social agent
    "asyncio==3.4.3",
    "asyncpg>=0.29.0", # project knowledge agent
    "boto3==1.37.18", # utils/crypto/crypto_nft.py
    "botocore==1.37.18", # utils/crypto/crypto_nft.py
    "duckduckgo-search==7.4.3", # ddg agent
    "fastapi==0.115.12", # for mesh api server
    "firecrawl-py==2.8.0", # firecrawl agent
    "flask==3.1.0",
    "flask[async]==3.1.0",
    "loguru==0.7.3",
    "mcp==1.6.0",
    "openai==1.71.0",
    "pre-commit==3.6.0",
    "prompt-toolkit==3.0.0",
    "psycopg2-binary==2.9.10", # core embeddings
    "py-cord==2.6.1", # interfaces/discord.py
    "pydash==8.0.5", # sol wallet agent
    "pyethash", # to build web3-ethereum-defi
    "pytest==7.4.3", # testing
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.0", # interfaces/telegram.py
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "ruff==0.11.8", # formatting/linting
    "safe-eth-py==6.0.0b42", # to build web3-ethereum-defi
    "scikit-learn==1.6.1", # for core embeddings
    "smolagents==1.9.2", # for coingecko, zkignite
    "spaceandtime==1.1.70", # for spaceandtime agent
    "tenacity==8.5.0", # sol wallet agent
    "tiktoken==0.9.0", # core embeddings
    "tweepy==4.15.0", # clients/twitter_api.py
    "typer[all]==0.9.0", # main_research.py
    "yfinance==0.2.54", # yahoo finance agent
    "uvicorn==0.34.0", # for mesh api server
    "web3-ethereum-defi==0.28.1", # aave agent
    "stockstats>=0.6.5",
    "google-cloud-bigquery>=3.27.0",
    "db-dtypes>=1.3.1",
    "undetected-chromedriver>=3.5.5",
    "google-genai>=1.56.0",
]

# required for spaceandtime agent as it uses old deps
[tool.uv]
override-dependencies = [
    "requests==2.32.3",
    "python-dotenv==1.1.0",
]

[tool.uv.sources]
pyethash = { git = "https://github.com/rexdotsh/ethash.git", rev = "master" } # forked version of ethash with windows support

[tool.ruff]
force-exclude = true
line-length = 120
lint.extend-select = ["I"]
